fix: connect image generation to card creation
Cards now have an @image_url attribute.
Templates can produce information about which field is to be used for image generation
Fields then use that to produce the text itself,
which is then passed to the existing OpenAiClient#get_image_url
Ultimately, AnkiConnectClient::DataService fetches a card's image_url when building the package to send to AnkiConnect.
Have just tested it and it appears to be working!